requirements.txt
pytesseract~=0.3.10
pillow~=10.3.0
openai~=1.23.6
streamlit~=1.33.0
pandas~=2.2.2
